requirements.txt
streamlit==1.31.0
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.1.9
langchain-community==0.0.24
langchain-groq==0.0.1
langchain-huggingface==0.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.1